cleanup old headers moved long ago
