--- PhysicsTools/PatUtils/interface/ElectronVPlusJetsIDSelectionFunctor.h
@@ -1,3 +1,0 @@
-#warning "This has moved to PhysicsTools/SelectorUtils"
-#include "PhysicsTools/SelectorUtils/interface/ElectronVPlusJetsIDSelectionFunctor.h"
-
