--- PhysicsTools/PatUtils/interface/MuonVPlusJetsIDSelectionFunctor.h
@@ -1,3 +1,0 @@
-#warning "This has moved to PhysicsTools/SelectorUtils"
-#include "PhysicsTools/SelectorUtils/interface/MuonVPlusJetsIDSelectionFunctor.h"
-
